Added: Plot

+++ Dev8-Scatterplots/src/main/java/com/jesseandswen/dev8/scatterplots/Models/DataModel.java
@@ -10,53 +10,31 @@ package com.jesseandswen.dev8.scatterplots.Models;
  * @author Jesse
  */
 public class DataModel {
-    
     private int CAT;
     private int EIG1;
-    private double EIG2;
+    private float EIG2;
 
-    /**
-     * @return the CAT
-     */
     public int getCAT() {
         return CAT;
     }
 
-    /**
-     * @param CAT the CAT to set
-     */
     public void setCAT(int CAT) {
         this.CAT = CAT;
     }
 
-    /**
-     * @return the EIG1
-     */
     public int getEIG1() {
         return EIG1;
     }
 
-    /**
-     * @param EIG1 the EIG1 to set
-     */
     public void setEIG1(int EIG1) {
         this.EIG1 = EIG1;
     }
 
-    /**
-     * @return the EIG2
-     */
-    public double getEIG2() {
+    public float getEIG2() {
         return EIG2;
     }
 
-    /**
-     * @param EIG2 the EIG2 to set
-     */
-    public void setEIG2(double EIG2) {
+    public void setEIG2(float EIG2) {
         this.EIG2 = EIG2;
     }
-    
-    
-    
 }
